pyproject.toml
[project]
name = "squeeze_evolve"
version = "0.1.0"
description = "Test-time scaling orchestration with confidence routing."
readme = "README.md"
requires-python = ">=3.10,<3.14"
authors = [{name = "SqueezeEvolve Contributors"}]
keywords = ["llm", "routing", "inference", "orchestration", "test-time-scaling"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Intended Audience :: Developers",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "fastapi>=0.115.0",
  "numpy>=1.26.0",
  "openai>=1.50.0",
  "pandas>=2.0.0",
  "pyarrow>=14.0.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.0",
  "tqdm>=4.66.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
aws = [
  "boto3>=1.34.0",
]
gcs = [
  "google-cloud-storage>=2.14.0",
]
scoring = [
  "openai-harmony",
  "transformers>=4.40.0",
]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.6.0",
]

[project.scripts]
squeeze-evolve-client = "squeeze_evolve.api.cli:client"
squeeze-evolve-server = "squeeze_evolve.api.cli:serve"
squeeze-evolve-eval = "squeeze_evolve.api.cli:client"

[tool.uv]
package = true

[tool.uv.sources]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/squeeze_evolve"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
